+++ src/main/java/kr/hhplus/be/server/domain/product/dto/ProductUpdateDto.java
@@ -3,5 +3,5 @@ package kr.hhplus.be.server.domain.product.dto;
 public record ProductUpdateDto(
         String name,
         Double price,
-        long stockAdjustment
+        int stockAdjustment
 ) {}
