+++ src/main/java/kr/hhplus/be/server/domain/product/entity/Stock.java
@@ -15,28 +15,28 @@ public class Stock extends BaseTimeEntity {
     private Long id;
 
     private Long productId;  // 상품 ID
-    private long quantity;   // 현재 재고 수량
+    private int quantity;   // 현재 재고 수량
 
     @Version
     private Long version;    // 낙관적 락을 위한 버전
 
-    public void decrease(long quantity) {
+    public void decrease(int quantity) {
         validateDecrease(quantity);
         this.quantity -= quantity;
     }
 
-    public void increase(long quantity) {
+    public void increase(int quantity) {
         validateIncrease(quantity);
         this.quantity += quantity;
     }
 
-    private void validateDecrease(long quantity) {
+    private void validateDecrease(int quantity) {
         if (this.quantity - quantity < 0) {
             throw new IllegalStateException("재고가 부족합니다");
         }
     }
 
-    private void validateIncrease(long quantity) {
+    private void validateIncrease(int quantity) {
         if (quantity <= 0) {
             throw new IllegalArgumentException("증가시킬 재고는 0보다 커야 합니다");
         }
